Refactor encode and decode. Add a `Name` abstraction.
The `encode` (née `code`) and `decode` functions now take buffer bounds,
and check them, so that they don't buffer-overflow.

Parsing domain names from the "hosts" file is newly robust. Backslash-escaping '.'
in a label is handled correctly. Encoding and decoding names is handled in a
single place, namely, the `encode` and `decode` methods of the `Name` class.
A `Name` is correctly represented as a sequence of labels, rather than as a
monolithic `std::string`.