Replace `Query` and `Response` with simply `Message`.
Support responses more complicated than just a single "answer" RR.
This seemed to be necessary in the short term just to fix the problem
that NOERROR responses were coming out malformed; so I replaced the
ad-hoc `Response::encode` with something that understood the actual
four-section format of a DNS response. This required adding classes
`Question` and `RR`; and then it seemed silly to keep `Query` and
`Response` as derived classes when they were basically the same thing.

The NOERROR responses given by this new server are *still* malformed
and crashy, but we're making progress.