Respect "zone cuts" during lookup and respond with referrals when appropriate.
This toy server puts all its RRs in a single physical file and in a
single physical database (tree) in memory; but the notion of "being
in an authoritative zone" is still meaningful. We are authoritative
if our path down through the domain tree has passed a top-of-zone
node (one with a SOA record) more recently than a zone-cut node (one
with an NS record but no SOA records).

Note that if a node contains a SOA record but no NS records, it is
ill-formed; we do not check for this during zone-file loading.

Note that if a zone cut occurs at a wildcard domain name (that is,
if there are NS records for the name `*.example.com`), then RFC 4592
says that "the semantics are left undefined." We do not check for
this situation during zone-file loading. If it occurs, nothing
surprising will happen.