Add decoding of "message compression" per RFC 1035.
This requires one of two possible approaches: either
(A) keep an exact copy of the original packet's bytes through
the whole process, so that compressed names can be decoded on the fly,
with appropriate sanity checking; or
(B) keep a pre-built "symbol table" through the whole process,
so that compressed names can be decoded in O(1) by consulting the table.

Here I go with approach (B), because it seems simpler and safer than (A).

The "compression" scheme defined in
https://tools.ietf.org/html/rfc1035
can express a 14-bit offset into the current packet. Any names represented
in this manner cannot be decoded without an appropriate `SymbolTable`.

My comments refer to, but do not implement, the "local compression"
scheme defined in
https://tools.ietf.org/html/draft-ietf-dnsind-local-compression-05
which can express a 14-bit offset into either the current RR's name or
the current RR's RDATA. This allows compressing RRs permanently (even
caching the compressed representation).