Start work on a simple "dig" client, a.k.a. stub resolver.
The program formerly known as "dnsserver" is now "dns-auth-server".
There is a new program, "dns-dig", which will soon behave like a
stub resolver:

    ./dns-auth-server 9000 zone.txt
    ./dns-dig 9000 www.google.com. ANY

However, for right now, `Digger::dig` is just an empty function;
it doesn't actually send the query over any network and it doesn't
receive or parse the response.

The output of `dns-dig` is modeled after `dig`, of course.