Turn RCode and Opcode into enum-like classes, not plain enums. NFC.
This lets me attach a `repr()` method to variables of these types,
so that for example I can write

    std::cout << RCode(3).repr() << std::endl;

instead of something like `rcode_repr(3)` or `repr(RCode(3))`.
The latter would actually be great, in my Python-flavored worldview,
but it is basically impossible to get that syntax in C++ because
if you are inside a function or method named `repr`, the one out in
the outer namespace will be shadowed (unless you `using ::repr`
explicitly, but that's horribly ugly and omission-prone).