Be more proper about FORMERR, NOTIMP, REFUSED, and blackholing queries.
We blackhole malformed packets, and also well-formed packets which are
not queries but rather (unsolicited) responses. Having trailing bytes
doesn't count as "malformed," at least not at the moment.

We return NOTIMP to well-formed queries whose opcode is not QUERY.

We return FORMERR to well-formed queries whose opcode is QUERY, but
which have specified no questions, or multiple questions. (See for
example https://stackoverflow.com/questions/4082081/requesting-a-and-aaaa-records-in-single-dns-query )

We return FORMERR if there is anything in the query's answer section,
or in its authority section.

We return FORMERR if there is anything in the query's additional section.
The additional section is expected to contain an RR of rrtype OPT if the
query is trying to use EDNS, but FORMERR is still the appropriate response
from servers which are not able to form a proper EDNS response. Example:

    dig @127.0.0.1 -p 9000 google.com. +dnssec

We return REFUSED to otherwise acceptable queries whose qname is not
a subdomain of any zone for which we are authoritative.

And incidentally, fix several embarrassing typo-bugs in Message::decode
and Message::encode.