Start working on a "bus" architecture for dns-dig.
The idea here is that we have one central "bus" that will take care
of sending out `Message` packets and receiving responses. Callers can
register to receive response packets (from a specific remote server,
on a specific port) by calling `Bus::listen`, which returns a
`future<Message>` representing the next packet to be received.
If nothing relevant is received in a given interval, then the future
should become ready with an exception of type `dns::TimeoutException`.