Implement zonefile syntax for unknown classes/types from RFC 3597.
And add some examples to "zone.txt". The one of type `TYPE16` is
a TXT record with the contents "hello" "world". (Notice that the
length of the entire RDATA is encoded in decimal, and then the
two string lengths, 5 and 5, are encoded a second time inside the
hexadecimal contents of the RDATA field.)